Update README.md
Learners_Academy_Project-1 Administrative-portal-for-Learner-s-Academy Project' Details This project aims to design and develop a backend administrative portal for the Learner’s Academy using Java EE technologies. I developed it as a project of phase 2 for the Become a back-end expert course. The goal of this project is to apply servlet, jsp and JDBC concepts.

Product Backlog: Create database and tables. Connect the database to the project. Create models classes. Create a database utility class to retrieve data. Create login page. Create JSP files for all pages of the project. Create a servlet to get requests and send responses to the JSP files. Add cookies. Create a CSS file to format the contents. Debug and Test the project. Technologies and tools Used • Servlet: to do the business logic and works a controller for the project.

• JSP: to handle the presentation view.

• SQL: to create and manage the database.

• JDBC: to make operations on the database for the project.

• CSS: to format the contents.

• phpMyAdmin: to administrate and manage the database manually.

• Eclipse: to write and run the code.

• Tomcat: to run and deploy servlet application.